manjiliao/excel_to_image/requirements.txt
dify_plugin>=0.7.0,<0.8.0
Spire.Xls.Free==14.12.4
plum-dispatch==1.7.4
